refactor: code split and restyle app service account dialogs
